+++ nri-redis/src/Redis.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# OPTIONS_GHC -fno-warn-redundant-constraints #-}
 
 -- | A simple Redis library providing high level access to Redis features we
 -- use here at NoRedInk
@@ -41,6 +42,13 @@ module Redis
     Internal.map2,
     Internal.map3,
     Internal.sequence,
+
+    -- * Every `Redis.Api key value` needs to implement an instance for `HasExamples value`.
+
+    -- | This instance is used to generate golden-tests for the api's value.
+    -- | You can use `Redis.Test.fromExamples yourRedisApi` in your test suite.
+    Examples.HasExamples (..),
+    Examples.example,
   )
 where
 
@@ -49,6 +57,7 @@ import qualified Data.ByteString as ByteString
 import Data.List.NonEmpty (NonEmpty)
 import qualified Data.List.NonEmpty as NonEmpty
 import qualified Dict
+import qualified Examples
 import qualified NonEmptyDict
 import qualified Redis.Codec as Codec
 import qualified Redis.Internal as Internal
@@ -132,15 +141,23 @@ data Api key a = Api
   }
 
 -- | Creates a json API mapping a 'key' to a json-encodable-decodable type
--- @
--- data Key = Key { fieldA: Text, fieldB: Text }
--- data Val = Val { ... }
 --
--- myJsonApi :: Redis.Api Key Val
--- myJsonApi = Redis.jsonApi (\Key {fieldA, fieldB}-> Text.join "-" [fieldA, fieldB, "v1"])
--- @
-jsonApi :: (Aeson.ToJSON a, Aeson.FromJSON a) => (key -> Text) -> Api key a
-jsonApi = makeApi Codec.jsonCodec
+-- > data Key = Key { fieldA: Text, fieldB: Text }
+-- > data Val = Val { ... }
+-- >
+-- > -- | This instance can be used to generate a golden test for this type.
+-- > -- | See Redis.Test
+-- > instance Redis.HasExamples Val where
+-- >   example = Examples.example "Val" Val { ... }
+-- >
+-- > myJsonApi :: Redis.Api Key Val
+-- > myJsonApi = Redis.jsonApi (\Key {fieldA, fieldB}-> Text.join "-" [fieldA, fieldB, "v1"])
+jsonApi ::
+  forall a key.
+  (Examples.HasExamples a, Aeson.ToJSON a, Aeson.FromJSON a) =>
+  (key -> Text) ->
+  Api key a
+jsonApi toKey = makeApi Codec.jsonCodec toKey
 
 -- | Creates a Redis API mapping a 'key' to Text
 textApi :: (key -> Text) -> Api key Text