+++ nri-redis/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased next version
 
+- In order to prevent accidental API changes, we force data in redis API to
+  implement the HasExamples TypeClass from the nri-test-encoding package. (thanks @stoeffel / #82)
+
 # 0.1.0.0
 
 - First release, but we've battle-tested it against significant load for months now!